requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
psycopg2
python-dotenv
requests
jinja2
aiofiles
python-multipart